zebra: add new flag to detect nexthop label updates
It doesn't make sense to set RIB_ENTRY_NEXTHOPS_CHANGED outside
zebra_rib.c because this flag is always cleared and reevaluated in the
rib_process() function.

With that said, define a new flag to indicate that the MPLS labels
of a nexthop have been changed. Then translate this flag into
RIB_ENTRY_NEXTHOPS_CHANGED in the nexthop_active_update() function.

This allows the NHT code to detect when labels are added/removed from
a nexthop and notify its clients accordingly.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>